陈三三 2017/12/05 内容模块(修改后)提交

+++ ft.serverone/src/main/java/com/future/technology/serverone/essay/dao/EssayMapper.java
@@ -23,7 +23,7 @@ public interface EssayMapper {
 
     int editorEssay(Essay essay);
 
-    int downline(Integer essay_id);
+    int downline(Long essay_id);
 
     List<EssayCustomer> queryEssay(QueryInfo queryInfo);
 }
