+++ ft.serverone/src/main/java/com/future/technology/serverone/essay/service/IEssayService.java
@@ -22,7 +22,7 @@ public interface IEssayService {
 
     Response updateEssay(Essay essay);
 
-    Response downline(Integer essay_id);
+    Response downline(Long essay_id);
 
     Response<PageBean> queryEssay(QueryInfo queryInfo);
